backend/requirements.txt
# ED-BASE Backend Requirements
# Production-grade dependencies for security-first Flask application

# Core Framework
Flask==3.0.0
gunicorn==21.2.0
Werkzeug==3.0.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Redis for rate limiting and caching
redis==5.0.1

# Supabase Auth integration
supabase==2.3.4
gotrue==2.3.0

# JWT handling
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# Stripe payments
stripe==7.8.0

# Security utilities
cryptography==41.0.7
bcrypt==4.1.2

# HTTP requests
httpx==0.26.0
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
email-validator==2.1.0

# Logging and monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Testing (dev)
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
